Add support for external bundle configuration.
This PR adds the ability to optionally specify multiple aggregations via an external configuration file.
This can be preferable over using multiple <execution> blocks as it allows other parts of the system and build pipeline
to understand resource bundling without being directly coupled to maven or this plugin.

This changeset is entirely backwards compatible. The Demo application has been updated to demonstrate this.