Remove order by filename.
Remove order by filename in class method "ProcessFilesTask.getFilesToInclude". The plugin should keep the original order to make it usable. In case of javascript files, user should put the files in the correct order before using the plugin to avoid javascript errors. The same case for css files, normally, the files are put on websites in order.